refactor : move navigation file , routes and main activity from app module to ui module
