+++ ui/src/main/java/com/cairosquad/ui/MainActivity.kt
@@ -1,11 +1,11 @@
-package com.cairosquad.movio
+package com.cairosquad.ui
 
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.activity.enableEdgeToEdge
 import com.cairosquad.design_system.theme.MovioTheme
-import com.cairosquad.movio.navigation.AppNavigation
+import com.cairosquad.ui.navigation.AppNavigation
 
 class MainActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
