+++ ui/src/main/java/com/cairosquad/ui/navigation/Routes.kt
@@ -1,4 +1,4 @@
-package com.cairosquad.movio.navigation
+package com.cairosquad.ui.navigation
 
 import kotlinx.serialization.Serializable
